requirements.txt
asgiref==3.2.3
Django==3.0.2
django-localflavor==2.2
django-phone-field==1.8.0
mysql-client==0.0.1
mysqlclient==1.4.6
python-stdnum==1.12
pytz==2019.3
sqlparse==0.3.0
